requirements.txt
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
gradio>=4.0.0
numpy>=1.24.0
